Fraud Recruitment Detector
This website is the analysis done by me regarding the fake job postings. I have used EMSCAD dataset for the evaluation.  The three classifier used by me are-
1. Random Forest Classifier.
2. KNN Classifier
3. Decision Tree Classifier
I have got the maximum accuracy of 98% from above 3 models. But if we compare the AUC I, can say that Random Forest and KNN performed good as compare to decision tree.